Moved acquire period tests out
